cardinallity 0 is working
