+++ fhirzeug/generators/python_pydantic/templates/resource.py.jinja2
@@ -36,7 +36,10 @@ class {{ clazz.name }}({{ clazz.superclass.name|default('object')}}):
 {% endfor %}
 
 {% for choice_prop, compound in clazz.choice_properties.items() %}
-    _{{choice_prop}}_choice_of_type_validator = pydantic.root_validator(allow_reuse=True)(choice_of_validator({{compound}}))
+#   {{ clazz.properties_map[compound[0]].__dict__}}
+
+    _{{choice_prop}}_choice_of_type_validator = pydantic.root_validator(allow_reuse=True) \
+     (choice_of_validator(set({{compound}}), {{clazz.properties_map[compound[0]].is_optional}}))
 {% endfor %}
 
     {# {% for prop in clazz.properties %} #}
