+++ fhirzeug/generators/python_pydantic/templates/resource_header.py
@@ -3,7 +3,7 @@ import enum
 import pydantic
 
 
-def choice_of_validator(choices):
+def choice_of_validator(choices, optional):
     def check_at_least_one(cls, values):
 
         setted_values = len(
@@ -11,7 +11,7 @@ def choice_of_validator(choices):
         )
         if setted_values > 1:
             raise ValueError(f"Only one of the fields is allowed to be set ({choices})")
-        elif setted_values < 1:
+        elif not optional and setted_values < 1:
             raise ValueError(f"A t least one of the fields needs to be set ({choices})")
         return values
